# backend/DTOs/UserDtos/UpdateUserDto.cs
using System.ComponentModel.DataAnnotations;
using backend.Constants;

namespace backend.DTOs.UserDtos;

public class UpdateUserDto
{
    [Required]
    [MaxLength(100)]
    public string FirstName { get; set; } = string.Empty;

    [Required]
    [MaxLength(100)]
    public string LastName { get; set; } = string.Empty;

    [Required]
    [MaxLength(150)]
    [EmailAddress]
    public string Email { get; set; } = string.Empty;

    [MaxLength(20)]
    [Phone]
    public string? ContactNumber { get; set; }

    [MaxLength(50)]
    public string UserRole { get; set; } = Roles.Customer;

    public bool IsActive { get; set; } = true;
}